finished splash and contact pages
