+++ src/App.js
@@ -11,6 +11,7 @@ import Projects from './pages/projects';
 import About from './pages/about';
 
 import ScrollToTopWrapper from './components/Common/ScrollToTopWrapper'
+import virtualportfolio from './pages/virtualportfolio';
 
 function App() {
   return (
@@ -20,9 +21,10 @@ function App() {
           <Route path="/" component={Home} exact/>
           <Route path="/signin" component={SignInPage} exact/>
           <Route path="/contact" component={Contact} exact/>
-          <Route path="/gallery" component={Photography} exact/>
+          <Route path="/photography" component={Photography} exact/>
           <Route path="/projects" component={Projects} exact/>
           <Route path="/about" component={About} exact/>
+          <Route path="/virtualportfolio" component={virtualportfolio} exact/>
         </Switch>
       </ScrollToTopWrapper>
