+++ src/components/Common/ButtonElements.js
@@ -3,11 +3,15 @@ import {Link as LinkR} from 'react-router-dom'
 import {Link as LinkS} from 'react-scroll'
 
 export const Button = styled(LinkS)`
+    --highlight: ${props => props.theme.highlight};
+    --foreground: ${props => props.theme.fg};
+    --background: ${props => props.theme.bg};
+
     border-radius: 50px;
-    background: ${({primary}) => (primary ? '#01bf71' : '#010606')};
+    background: ${({lightbg}) => (lightbg ? 'var(--foreground)' : 'var(--background)')};
     white-space: nowrap;
     padding: ${({big}) => (big ? '14px 48px' : '12px 30px')};
-    color: ${({dark}) => (dark ? '#010606' : '#fff')};
+    color: ${({lightbg}) => (!lightbg ? 'var(--background)' : 'var(--foreground)')};
     font-size: ${({fontbig}) => (fontbig ? '20px' : '16px')};
     outline: none;
     border: none;
@@ -19,16 +23,22 @@ export const Button = styled(LinkS)`
 
     &:hover {
         transition: all 0.2 ease-in-out;
-        background: ${({primary}) => (primary ? '#fff' : '#01bf71')};
+        background: ${({lightbg}) => (lightbg ? 'var(--foreground)' : 'var(--background)')};
     }
 `
 
 export const RouteButton = styled(LinkR)`
+    --highlight: ${props => props.theme.highlight};
+    --foreground: ${props => props.theme.fg};
+    --background: ${props => props.theme.bg};
+    --text2: ${props => props.theme.t1};
+    --text: ${props => props.theme.t3};
     border-radius: 50px;
-    background: ${({primary}) => (primary ? '#01bf71' : '#010606')};
+    /* background: ${({lightbg}) => (lightbg ? 'var(--foreground)' : 'var(--background)')}; */
+    background: var(--highlight);
     white-space: nowrap;
     padding: ${({big}) => (big ? '14px 48px' : '12px 30px')};
-    color: ${({dark}) => (dark ? '#010606' : '#fff')};
+    color: ${({lightbg}) => (!lightbg ? 'var(--text)' : 'var(--text2)')};
     font-size: ${({fontbig}) => (fontbig ? '20px' : '16px')};
     outline: none;
     border: none;
@@ -40,6 +50,6 @@ export const RouteButton = styled(LinkR)`
 
     &:hover {
         transition: all 0.2 ease-in-out;
-        background: ${({primary}) => (primary ? '#fff' : '#01bf71')};
+        background: ${({lightbg}) => (lightbg ? 'var(--background)' : 'var(--foreground)')};
     }
 `
