+++ src/components/Contact/ContactForm/ContactFormElements.js
@@ -15,29 +15,34 @@ export const FormPageContainer = styled.div`
 `
 //Contains Video and shadow effect
 export const FormBackground = styled.div`
-    /* border: 1px solid red; */
+    border: 1px solid red;
     transform: skewY(-11deg);
     position: absolute;
-    top: -1px;
+    top: 0px;
     right: 0;
     bottom: 0;
     left: 0;
     width: 100%;
     /* height: 100%; */
+    margin-bottom: 10px;
     overflow: hidden;
+
+    @media screen and (min-width: 2000px){
+        margin-bottom: 40px;
+    }
 `
 
-//Motion div for moving video
-export const DiagonalBox = styled(motion.div)`
-    position: absolute;
-    right: 0;
-    left: 0;
-    bottom: 0;
-    background: #fff;
-    height: 300px;
-    box-shadow: inset 0px 12px 12px -9px ${props => props.theme.shadow}, inset 0px -12px 12px -7px ${props => props.theme.shadow};
-    z-index:2;
-`
+// //Motion div for moving video
+// export const DiagonalBox = styled.div`
+//     position: absolute;
+//     right: 0;
+//     left: 0;
+//     bottom: 0;
+//     background: #fff;
+//     height: 300px;
+//     box-shadow: inset 0px 12px 12px -9px ${props => props.theme.shadow}, inset 0px -12px 12px -7px ${props => props.theme.shadow};
+//     z-index:2;
+// `
 
 //Contains All Forground elements
 export const PageWrapper = styled.div`
@@ -95,20 +100,38 @@ export const InputContainer = styled.div`
 /**
  * Content
  */
-export const VideoBackground = styled(motion.video)`
-    border: 1px solid green;
+export const VideoBackground = styled.video`
+    /* border: 1px solid green; */
 
-    transform: skewY(11deg);
-    width: 100%;
-    position: absolute;
-    /* top: -200px; */
+    /* transform: skewY(11deg); */
+    /* height: 130vh; */
+    /* position: absolute; */
+    /* top: -1000px;
     right: 0;
     bottom: 0;
-    left: 0;
+    left: 0; */
     /* height: 200%; */
     -o-object-fit: cover;
     object-fit: cover;
+    background-size: cover;
+
     z-index:1;
+    
+    right: 0; 
+    bottom: 0;
+    min-width: 100%; 
+    min-height: 100%;
+    width: auto; 
+    height: auto; 
+`
+
+export const VideoOffsetter = styled(motion.div)`
+    transform: skewY(11deg);
+    /* border: 3px solid pink; */
+    height: 100%;
+    width: 100%;
+    position: absolute;
+    top: -200px;
 `
 
 export const ShadowBox = styled.div`
