+++ src/components/Home/theme.js
@@ -1,16 +1,16 @@
 export const theme = {
-    fg: "#fff",
-    bg: "#000",
-    fg1: "#eee",
-    bg1: "#222",
-    t1: "#fff",
-    t2: "#595959",
-    t3: "#fff",
-    t4: "#ddd",
+    fg: "#ddd", //Light fg
+    bg: "#000", //Dark bg
+    fg1: "#eee", //Dark fg
+    bg1: "#222", //Light fg
+    t1: "#fff", //Light text
+    t2: "#ddd", //Off-Light Text
+    t3: "#000", //Dark Text
+    t4: "#333", //Off- Dark Text
     textField: "#eee",
     border: "#737373",
     active: "#c74a5a",
-    focus: "#8a0011",
-    highlight: "#c74a5a",
+    focus: "#01bf71",
+    highlight: "#01bf71", //Page color
     shadow: "#000"
 };
